fix build errors

+++ onnxruntime/test/providers/cpu/llm/attention_op_test.cc
@@ -220,10 +220,10 @@ static void RunTest3D(
   std::vector<int64_t> v_shape = {batch_size, kv_sequence_length, v_hidden_size};
 
   std::vector<int64_t> attn_mask_shape = {q_sequence_length, total_sequence_length};
-  if (q_sequence_length * total_sequence_length != attn_mask.size() && attn_mask.size() > 0) {
-    if (batch_size * q_sequence_length * total_sequence_length == attn_mask.size()) {
+  if (q_sequence_length * total_sequence_length != static_cast<int>(attn_mask.size()) && attn_mask.size() > 0) {
+    if (batch_size * q_sequence_length * total_sequence_length == static_cast<int>(attn_mask.size())) {
       attn_mask_shape = {batch_size, 1, q_sequence_length, total_sequence_length};
-    } else if (batch_size * q_num_heads * q_sequence_length * total_sequence_length == attn_mask.size()) {
+    } else if (batch_size * q_num_heads * q_sequence_length * total_sequence_length == static_cast<int>(attn_mask.size())) {
       attn_mask_shape = {batch_size, q_num_heads, q_sequence_length, total_sequence_length};
     } else {
       ORT_THROW("Invalid attention mask size: ", attn_mask.size(),
@@ -298,10 +298,10 @@ static void RunTest4D(
   std::vector<int64_t> v_shape = {batch_size, kv_num_heads, kv_sequence_length, v_head_size};
 
   std::vector<int64_t> attn_mask_shape = {q_sequence_length, total_sequence_length};
-  if (q_sequence_length * total_sequence_length != attn_mask.size() && attn_mask.size() > 0) {
-    if (batch_size * q_sequence_length * total_sequence_length == attn_mask.size()) {
+  if (q_sequence_length * total_sequence_length != static_cast<int>(attn_mask.size()) && attn_mask.size() > 0) {
+    if (batch_size * q_sequence_length * total_sequence_length == static_cast<int>(attn_mask.size())) {
       attn_mask_shape = {batch_size, 1, q_sequence_length, total_sequence_length};
-    } else if (batch_size * q_num_heads * q_sequence_length * total_sequence_length == attn_mask.size()) {
+    } else if (batch_size * q_num_heads * q_sequence_length * total_sequence_length == static_cast<int>(attn_mask.size())) {
       attn_mask_shape = {batch_size, q_num_heads, q_sequence_length, total_sequence_length};
     } else {
       ORT_THROW("Invalid attention mask size: ", attn_mask.size(),
